+++ backend/utils/dependencies.py
@@ -1,4 +1,4 @@
-from fastapi import Depends, HTTPException, status
+from fastapi import Depends, HTTPException, status, Cookie
 from fastapi.security import HTTPBearer, HTTPAuthorizationCredentials
 from jose import JWTError, jwt
 from sqlalchemy.orm import Session
@@ -9,7 +9,6 @@ from utils.auth import SECRET_KEY, ALGORITHM
 bearer_scheme = HTTPBearer(auto_error=False) 
 def get_current_user(credentials: HTTPAuthorizationCredentials = Depends(bearer_scheme), db: Session = Depends(get_db)):
     if not credentials:
-        print("❌ No credentials")
         raise HTTPException(status_code=403)
 
     if credentials.scheme.lower() != "bearer":
@@ -19,16 +18,18 @@ def get_current_user(credentials: HTTPAuthorizationCredentials = Depends(bearer_
         payload = jwt.decode(credentials.credentials, SECRET_KEY, algorithms=[ALGORITHM])
         email = payload.get("sub")
         if not email:
-            print("❌ No sub in payload")
             raise HTTPException(status_code=403)
     except JWTError as e:
-        print("❌ JWT error:", e)
         raise HTTPException(status_code=403)
 
     user = db.query(UserModel).filter(UserModel.email == email).first()
-    print("🧍 Found user:", user)
     if user is None:
-        print("❌ No user in DB")
         raise HTTPException(status_code=403)
 
     return user
+
+def get_current_refresh_token(refresh_token: str = Cookie(None)):
+    if not refresh_token:
+        raise HTTPException(status_code=401, detail="Refresh token missing")
+    return refresh_token
+